Matt refactor
